requirements.txt
SQLAlchemy==1.4.13
alembic==1.6.0
fastapi==0.63.0
uvicorn==0.13.4
psycopg2-binary==2.8.6
FastAPI-SQLAlchemy
python-dotenv
babel
pydantic
passlib
PyJWT
bcrypt
pydantic[email]
google-auth
requests
fastapi-mail==0.3.7
aiofiles
unidecode